feat(app): floating recording playback sheet (replaces full-page detail)
A floating bottom sheet (transparent modal + inset rounded card) with
waveform, transport, Sync now, and share/details/delete — on the app's
white/black palette (white primary CTA, neutral controls, red delete).
